Add a function to unpack previously-packed values.
transform_out is the opposite of transform_in, and unpack is the opposite
of pack. Given that transform_in is somewhat liberal in what it accepts
this is not completely symmetrical, but assuming the caller is passing
in values of the canonical type (i.e. unicode values into string fields)
the unpacked value should match what was packed.