Fix TREE_MODE to be the right value, causing better behavior.
There was one too many zeros in the file mode gitdict was using for trees,
which was what was causing tree_changes to fail to visit the child nodes.
This means find_change_leaves needs to be reimagined a bit, but for now
this just reduces its functionality a bit until a later point where it
can be rewritten to behave more like it used to behave.